Added transition to fixed style after animation
This seems to fix jankiness in title position when scrolling through
pages, and provides a more robust method to format the title position in
the top bar. There is a small offset in the text alignment between
firefox and chrome which I've been unable to fix, just found a middle
point that is acceptable in each.
Next steps are anchoring scroll around each page, making top buttons
function, and adding scroll-based indicator to show which page is
active.