Visual scroll and page indicator on nav buttons
Added onclick effects for the buttons on the top navigation bar to now
scroll to the corresponding section. Implemented a custom function for
this to be able to customise the scroll speed -- set fairly slow, we're
not in a rush. Also clicking the title below the nav buttons now brings
you to the title page.

Added an 'indicator' to show which page is currently active. Indicator
simply scrolls across the top bar as the user scrolls down the page.

Fairly happy with this as a basic skeleton, ready to fill in the details
on each page. One step further would be to hijack the user's scroll to
only switch between pages. This will be complicated to implement and
probably not worth the time given what I've done so far with
scroll-linked effects -- I like what I have here. The compromise might
be to see if I can temporarily lock the scroll to the bottom of the page
to make it easier to stop on a whole page. This would be an extension
once the content is complete however.