Added infrastructure for new 'InfoBlock's
Added InfoBlocks as a generic way to display information on projects and
skills, will form a bunch of objects that can go into a flexbox container.

Got stuck with some CSS trying to make my research projects div overflow before
the whole page overflows, without removing the possibility for the page to
overflow. So far the only solution I found is fixing the height of the body in
JS, which I'm not happy with as it's too hacky. May need to add a fixed space
for the page header but then it might split onto multiple lines... not sure
what the soln is yet.