Updated styles
Playing with overall styling, added colour variables in CSS. Starting to look a
bit more put-together now which is nice. Will probably want to play with font
and more colour stuff in future.

Re: the problem mentioned in the last commit, I implemented the JS hack as I
couldn't find a way around it.