Updating information and developing design
Updated some of the information, styling, images in the about and
education pages. Added some infrastructure for skills page.

Want to overhaul the interface for displaying the 'infoblocks' set up in
JS code. Plan to have each page with a list of items and summary on the
left or right, with a panel on the other half of the page that gives
further detail on the selected item. This will need an alternative
implementation for mobile aspect rations, not sure what this will be
yet.